+++ bust-a-move-framework/src/main/java/nl/tudelft/bust_a_move20162017/bust_a_move_framework/cannon/Cannon.java
@@ -150,12 +150,11 @@ public class Cannon {
      */
     private Bubble getNextBubble() {
         game.log.log(this, "Next bubble loaded to cannon");
-        Bubble newBubble = Bubble.randomColor((double) this.xLaunch,
-                (double) this.yLaunch, game.arena.getBubbleStorage().getColorsOnArena(), true);
+        Bubble newBubble = Bubble.randomColor((float) this.xLaunch,
+                (float) this.yLaunch, game.arena.getBubbleStorage().getColorsOnArena(), true);
         if (GameConfig.ENABLE_POWERUPS) {
             newBubble = PowerUp.apply(newBubble);
         }
-
         return newBubble;
     }
     /**
